requirements.txt
selenium
numpy
flask
pyvirtualdisplay
Pillow
boto3
fake-useragent
requests
psycopg2-binary